Improve TensorFlow Device Placement
This updates the model compilation logic to intelligently colocate ops
for targets, metrics, and losses on the same device as the associated
model output that is consumed by these ops.

For example, this would allow a user to create a multi-input,
multi-output, shared model to efficiently exploit multiple GPUS.
I.e. if a user defines a shared model on the CPU, creates "towers"
for each GPU with separate inputs and outputs, and then creates a
multi-input, multi-output model to operate on split batches, this
fix will cause the associated targets, metrics, and losses for each
output to be located on the same device as the output, rather than
being placed on the default device.  Otherwise, the GPUs will run the
ops for their associated tower, and then send their results the default
device (first GPU) to run the ops for targets, metrics, and losses,
thus wasting the resources of the remaining GPUs.

Improvement on previous dual-GPU setup (batch_size=32, Tesla K80 GPU
(which effectiely contains 2 GPUs), 100 batches): ~1.4x

Total improvement for dual-GPU setup vs. single-GPU setup (batch_size=32,
Tesla K80 GPU (which effectiely contains 2 GPUs), 100 batches): ~1.7-1.8x