Improve TensorFlow Device Placement
This updates the model compilation logic to colocate ops for targets,
metrics, losses, and sample weights on the same device as the associated
model output that is consumed by these ops.

For example, this would allow a user to create a multi-input,
multi-output, shared model to efficiently exploit multiple GPUs.
I.e. if a user defines a shared model on the CPU, creates "towers"
for each GPU with separate inputs and outputs, and then creates a
multi-input, multi-output model to operate on split batches, this
fix will cause the associated targets, metrics, and losses for each
output to be located on the same device as the output, rather than
being placed on the default device.  Otherwise, the GPUs will run the
ops for their associated tower, and then send their results the default
device (first GPU) to run the ops for targets, metrics, losses, and
sample weights, thus wasting the resources of the remaining GPUs.